// client/src/redux/personnel/personnel.types.ts
export enum ActionType {
    PERSONNEL_TABLE_LOAD_USER_REQUEST = 'PERSONNEL_TABLE_LOAD_USER_REQUEST',
    PERSONNEL_TABLE_LOAD_USER_SUCCESS = 'PERSONNEL_TABLE_LOAD_USER_SUCCESS',
    PERSONNEL_TABLE_LOAD_USER_FAILED = 'PERSONNEL_TABLE_LOAD_USER_FAILED',


    PERSONNEL_TABLE_ADD_USER_REQUEST = 'PERSONNEL_TABLE_ADD_USER_REQUEST',
    PERSONNEL_TABLE_ADD_USER_SUCCESS = 'PERSONNEL_TABLE_ADD_USER_SUCCESS',
    PERSONNEL_TABLE_ADD_USER_FAILED = 'PERSONNEL_TABLE_ADD_USER_FAILED',
    PERSONNEL_TABLE_ADD_USER_RESET = 'PERSONNEL_TABLE_ADD_USER_RESET',


    PERSONNEL_TABLE_DELETE_USER_REQUEST = 'PERSONNEL_TABLE_DELETE_USER_REQUEST',
    PERSONNEL_TABLE_DELETE_USER_SUCCESS = 'PERSONNEL_TABLE_DELETE_USER_SUCCESS',
    PERSONNEL_TABLE_DELETE_USER_FAILED = 'PERSONNEL_TABLE_DELETE_USER_FAILED',


    PERSONNEL_TABLE_UPDATE_USER_REQUEST = 'PERSONNEL_TABLE_UPDATE_USER_REQUEST',
    PERSONNEL_TABLE_UPDATE_USER_SUCCESS = 'PERSONNEL_TABLE_UPDATE_USER_SUCCESS',
    PERSONNEL_TABLE_UPDATE_COMPLETE = 'PERSONNEL_TABLE_UPDATE_COMPLETE',
    PERSONNEL_TABLE_UPDATE_USER_RESET = 'PERSONNEL_TABLE_UPDATE_USER_RESET',


    PERSONNEL_TABLE_FAILED = 'PERSONNEL_TABLE_FAILED',


    PERSONNEL_TABLE_LOAD_COMPANIES_SUCCESS = 'PERSONNEL_TABLE_LOAD_COMPANIES_SUCCESS',
    PERSONNEL_TABLE_LOAD_COMPANIES_AND_PERSONNEL_SUCCESS = 'PERSONNEL_TABLE_LOAD_COMPANIES_AND_PERSONNEL_SUCCESS'
}